Refactor django blocks derived classes into seperate namespace and folder.
Also, delete some unused using statements
