+++ Python/Product/Django/Django.csproj
@@ -205,7 +205,20 @@
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="PkgCmdID.cs" />
     <Compile Include="TemplateContentType.cs" />
+    <Compile Include="TemplateParsing\DjangoBlocks\BlockParseInfo.cs" />
+    <Compile Include="TemplateParsing\DjangoBlocks\DjangoArgumentlessBlock.cs" />
+    <Compile Include="TemplateParsing\DjangoBlocks\DjangoAutoEscapeBlock.cs" />
     <Compile Include="TemplateParsing\DjangoBlock.cs" />
+    <Compile Include="TemplateParsing\DjangoBlocks\DjangoFilterBlock.cs" />
+    <Compile Include="TemplateParsing\DjangoBlocks\DjangoForBlock.cs" />
+    <Compile Include="TemplateParsing\DjangoBlocks\DjangoIfBlock.cs" />
+    <Compile Include="TemplateParsing\DjangoBlocks\DjangoIfOrIfNotEqualBlock.cs" />
+    <Compile Include="TemplateParsing\DjangoBlocks\DjangoLoadBlock.cs" />
+    <Compile Include="TemplateParsing\DjangoBlocks\DjangoMultiVariableArgumentBlock.cs" />
+    <Compile Include="TemplateParsing\DjangoBlocks\DjangoSpacelessBlock.cs" />
+    <Compile Include="TemplateParsing\DjangoBlocks\DjangoTemplateTagBlock.cs" />
+    <Compile Include="TemplateParsing\DjangoBlocks\DjangoUnknownBlock.cs" />
+    <Compile Include="TemplateParsing\DjangoBlocks\DjangoWidthRatioBlock.cs" />
     <Compile Include="TemplateParsing\DjangoFilter.cs" />
     <Compile Include="TemplateParsing\DjangoFormatDefinitions.cs" />
     <Compile Include="TemplateParsing\DjangoVariable.cs" />
